fix: don't redirect with slashes
In FastAPI, if the resource is {uri}/resource/ and we do {uri}/resource,
the library will return a 307 redirect to {uri}/resource/.

So, this means another network call, which means app/web is slower.

Plus, it redirects with http, not https, which cause CORS to go mad.

More info in this issue:
https://github.com/tiangolo/fastapi/issues/2060#issuecomment-834868906